Adding BambooAgent with agent 2.0 configuraion for hari.
