+++ PlatformAgents/com/cognizant/devops/platformagents/agents/ci/bamboo/InSightsBambooAgent.service
@@ -1,0 +1,10 @@
+[Unit]
+Description=Ubuntu CITFSAgent Service
+
+[Service]
+Type=forking
+ExecStart=/bin/bash /opt/insightsagents/PlatformAgents/citfs/InSightsCITFSAgent.sh start
+ExecStop=/bin/bash /opt/insightsagents/PlatformAgents/citfs/InSightsCITFSAgent.sh stop
+
+[Install]
+WantedBy=multi-user.target
